feat: make projection more similar to standard basis
This sorts the basis1 and basis2 vectors so that XY, XZ, and YZ
vectors appear in that order. It also checks the direciton of
the normal and reflects bases into the positive or negative zone
depending on the way which way it points (towards or away from
<1,1,1>).

This gives a somewhat sensible behavior where <0,0,1> gives
and XY plane in the expected orientation and <0,0,-1> gives
it in a reflected orientation. Likewise for ihat, and jhat.